add back in tests that were commented out
