+++ src/IdentityServer4/Configuration/DependencyInjection/BuilderExtensions/Core.cs
@@ -108,7 +108,7 @@ namespace Microsoft.Extensions.DependencyInjection
             builder.Services.AddTransient<IdentityServerTools>();
 
             builder.Services.AddTransient<IReturnUrlParser, OidcReturnUrlParser>();
-            builder.Services.AddTransient<IUserSession, CookieAuthenticationUserSession>();
+            builder.Services.AddTransient<IUserSession, DefaultUserSession>();
             builder.Services.AddTransient(typeof(MessageCookie<>));
             builder.Services.AddScoped<AuthenticationHandler>();
