+++ src/IdentityServer4/Endpoints/Results/CheckSessionResult.cs
@@ -6,7 +6,6 @@ using System.Threading.Tasks;
 using IdentityServer4.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
-using IdentityServer4.Services;
 using IdentityServer4.Configuration;
 
 namespace IdentityServer4.Endpoints.Results
