+++ test/IdentityServer.UnitTests/Common/MockClientSessionService.cs
@@ -2,9 +2,6 @@
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
 
-using IdentityServer4.Services;
-using System.Collections.Generic;
-using System.Threading.Tasks;
 
 namespace IdentityServer4.UnitTests.Common
 {
