+++ test/IdentityServer.UnitTests/Common/MockSessionIdService.cs
@@ -2,9 +2,6 @@
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
 
-using IdentityServer4.Services;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http.Features.Authentication;
 
 namespace IdentityServer4.UnitTests.Common
 {
